/trunk, vctPlot2DOpenGL: removed unused variable
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk/saw/components/sawRobotIO1394@4324 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ include/sawRobotIO1394/mtsRobotIO1394QtWidget.h
@@ -126,7 +126,7 @@ protected:
 private:
     mtsIntervalStatistics IntervalStatistics;
 
-    int NumberOfActuators;
+    size_t NumberOfActuators;
 
     vctDoubleVec UnitFactor;
     vctDoubleVec JointPosition;
